add GATK jar file
The Broad Office of Strategic Alliances and Partnering has given the
green light for public (re)distribution of GATK3, in line with how
GATK4 will be licensed and distributed. This commit causes the
GenomeAnalysisTK.jar file to be included in the conda package, sourced
from the now-unauthenticated URL for GATK. Note: While older GATK
versions can be accessed via URLs that include the version number, it
seems the latest version can only be pulled from a URL that omits the
version.